services/data_consuming/requirements.txt
kafka-python==2.2.15
pymongo==4.14.1
elasticsearch==8.11.0
